GCP_projects/SPA-and-API/backend/requirements.txt
cloud-sql-python-connector==1.8.0
Flask==3.0.2
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
flask-talisman==1.1.0
google-cloud-secret-manager==2.20.1
gunicorn==21.2.0
pg8000==1.30.5
